requirements.txt
streamlit>=1.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.2.0
lxml>=4.9.0
protobuf==3.20.3 